web_server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
psycopg2-binary>=2.9.9
qdrant-client>=1.8.0
numpy>=1.26.0
openai>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
